Roles Authorization implement

+++ database/migrations/2025_07_16_204259_remove_role_column_from_users_table.php
@@ -5,12 +5,15 @@ use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration {
-    public function up(): void
-    {
-        Schema::table('users', function (Blueprint $table) {
+public function up()
+{
+    Schema::table('users', function (Blueprint $table) {
+        if (Schema::hasColumn('users', 'role')) {
             $table->dropColumn('role');
-        });
-    }
+        }
+    });
+}
+
 
     public function down(): void
     {
